part 4 tasks 4.15-4.23

+++ controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const middleware = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -17,23 +18,26 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const body = request.body
+blogsRouter.post('/', middleware.userExtractor, async (request, response, next) => {
+  try {
+    const body = request.body  // Ei tarvitse käyttää await, koska body on jo saatavilla
+    const user = request.user  // Sama juttu, user on asetettu middlewaresta
 
-  const user = await User.findById(body.userId)
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      user: user._id,
+    })
 
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    user: user._id,
-  })
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
 
-  const savedBlog = await blog.save()
-  user.notes = user.notes.concat(savedBlog._id)
-  await user.save()
-
-  response.status(201).json(savedBlog)
+    response.status(201).json(savedBlog)
+  } catch (error) {
+    next(error) // Siirretään virhe error handler -middlewarelle
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -66,9 +70,22 @@ blogsRouter.put('/:id', async (request, response) => {
   response.json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+
+  const userId = await request.user.id.toString()
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  const blogAdderId = blog.user.toString()
+
+  if ( userId ===   blogAdderId ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  } else {
+    return  response.status(403).json({ error: 'No permission to delete this blog' })
+  }
 })
 
 module.exports = blogsRouter
